+++ src/components/form/Select.tsx
@@ -1,1 +1,8 @@
-export { Select } from "./Select.styled";
+import type { ComponentProps } from "react";
+
+import { Select } from "./Select.styled";
+
+type SelectProps = ComponentProps<typeof Select>;
+
+export type { SelectProps };
+export { Select };